Move utility helper functions from within extractor.py and download.py to the utilities folder in database_queries.py or date_conversions.py.
